+++ store/src/apis/utils/index.ts
@@ -53,7 +53,7 @@ const axiosAuthApi = (baseURL: string | undefined) => {
               window.location.href = "/login";
             });
           return axios(originalRequest);
-        } else {
+        } else if (error.response.data.message === "Refresh-Expired") {
           localStorage.clear();
           // eslint-disable-next-line no-restricted-globals
           window.location.href = "/login";
